Adds a bunch of integration tests for the GameController class

+++ src/main/kotlin/com/commerzbank/VideoGameDatabase/service/GameService.kt
@@ -6,7 +6,9 @@ import com.commerzbank.VideoGameDatabase.dto.GameDto
 import com.commerzbank.VideoGameDatabase.logger.Log
 import com.commerzbank.VideoGameDatabase.model.Game
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.dao.EmptyResultDataAccessException
 import org.springframework.stereotype.Service
+import javax.persistence.EntityNotFoundException
 import javax.transaction.Transactional
 import javax.xml.ws.ServiceMode
 import kotlin.math.log
@@ -24,13 +26,10 @@ open class GameService {
     }
 
     @Transactional
-    open fun get(id: Int): GameDto{
-        return try{
-            GameDto(dao.getOne(id))
-        } catch (e: Exception){
-            Log.log.error(e.message)
-            GameDto()
-        }
+    open fun get(id: Int): GameDto?{
+        if(dao.existsById(id))
+            return GameDto(dao.getOne(id))
+        return null
     }
 
     @Transactional
@@ -40,6 +39,22 @@ open class GameService {
         }
     }
 
+    @Transactional
+    open fun paginate(page:Int,count:Int):List<GameDto>{
+        val games = listAll()
+        val list = mutableListOf<GameDto>()
+        val index = (page-1)*count
+        for (x in index until page*count){
+            if(x > games.size-1){
+                return list
+            }
+            else{
+                list.add(games[x])
+            }
+        }
+        Log.log.info(list.size.toString())
+        return list
+    }
 
     @Transactional
     open fun create(dto: GameDto){
@@ -48,18 +63,31 @@ open class GameService {
 
 
     @Transactional
-    open fun update(dto: GameDto){
-        val game = dao.getOne(dto.id!!)
-        game.name = dto.name
-        game.price = dto.price
-        game.publisher = dto.publisher
-        game.rating = dto.rating
-        dao.save(game)
+    open fun update(dto: GameDto):GameDto?{
+        val id = dto.id!!
+        return if (dao.existsById(id)){
+            val game = dao.getOne(id)
+            game.name = dto.name
+            game.price = dto.price
+            game.publisher = dto.publisher
+            game.rating = dto.rating
+            dao.save(game)
+            dto
+        } else{
+            null
+        }
     }
 
     @Transactional
-    open fun delete(id: Int){
-        dao.deleteById(id)
+    open fun delete(id: Int):GameDto?{
+        if(dao.existsById(id)){
+            val game = dao.getOne(id)
+            dao.deleteById(id)
+            return GameDto(game)
+        }
+        return null
+
+
     }
 
     @Transactional
@@ -85,6 +113,15 @@ open class GameService {
         return dao.findByPublisher(publisher).map { GameDto(it) }
     }
 
+    @Transactional
+    open fun exists(id:Int):Boolean{
+        return dao.existsById(id)
+    }
+
+    @Transactional
+    open fun count(): Long {
+        return dao.count()
+    }
 
 
 }
